backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
pydantic==2.7.0
pydantic-settings==2.3.0
scikit-learn==1.5.0
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
pyserial==3.5
httpx==0.27.0
openai==1.35.0
python-dotenv==1.0.1
